+++ driver/pom.xml
@@ -28,6 +28,7 @@
     <properties>
         <spray.version>1.3.3</spray.version>
         <akka.version>2.3.4-spark</akka.version>
+        <spray.test.version>1.3.1</spray.test.version>
     </properties>
     <dependencies>
         <dependency>
@@ -143,37 +144,37 @@
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-can</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-http</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-httpx</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-util</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-client</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>io.spray</groupId>
             <artifactId>spray-json_2.10</artifactId>
-            <version>1.3.1</version>
+            <version>${spray.test.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -185,8 +186,8 @@
         <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>casbah-core_${scala.binary.version}</artifactId>
-            <scope>test</scope>
             <version>2.8.0</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
